chore: move slug to url.js
