+++ packages/rum-core/src/performance-monitoring/performance-monitoring.js
@@ -30,7 +30,7 @@ import {
   stripQueryStringFromUrl,
   getDtHeaderValue
 } from '../common/utils'
-import Url from '../common/url'
+import { Url } from '../common/url'
 import { patchEventHandler } from '../common/patching'
 import { globalState } from '../common/patching/patch-utils'
 import {
