+++ packages/rum-core/src/performance-monitoring/transaction-service.js
@@ -48,7 +48,7 @@ import {
 } from '../common/constants'
 import { addTransactionContext } from '../common/context'
 import { __DEV__, state } from '../state'
-import slugifyUrl from '../common/slugify'
+import { slugifyUrl } from '../common/url'
 
 class TransactionService {
   constructor(logger, config) {
@@ -233,6 +233,12 @@ class TransactionService {
      */
     this.recorder.stop()
 
+    /**
+     * Capturing it here before scheduling the transaction end
+     * as to avoid capture different location when routed
+     */
+    const currentUrl = window.location.href
+
     return Promise.resolve().then(
       () => {
         const { name, type } = tr
@@ -279,7 +285,7 @@ class TransactionService {
          * Categorize the transaction based on the current location
          */
         if (tr.name === NAME_UNKNOWN) {
-          tr.name = slugifyUrl(window.location.href)
+          tr.name = slugifyUrl(currentUrl)
         }
 
         captureNavigation(tr)
