+++ packages/rum-core/test/common/url.spec.js
@@ -23,7 +23,7 @@
  *
  */
 
-import Url from '../../src/common/url'
+import { Url, slugifyUrl } from '../../src/common/url'
 
 describe('Url parser', function() {
   it('should parse relative url', function() {
@@ -363,3 +363,45 @@ if (window.URL.toString().indexOf('native code') !== -1) {
     }
   })
 }
+
+describe('Slug URL', () => {
+  const validate = (before, after, depth) => {
+    expect(slugifyUrl(before, depth)).toBe(after)
+  }
+
+  it('accept depth parameter', () => {
+    validate('/a/b/c/d', '/a/b/*')
+    validate('/a/b/c/d', '/a/*', 1)
+  })
+
+  it('handle trailing slash', () => {
+    validate('/a/', '/a')
+    validate('/a', '/a')
+    validate('/', '/')
+  })
+
+  it('handle query param', () => {
+    validate('/a/b/?id=hello', '/a/b?{query}')
+    validate('/?foo=bar&bar=baz', '/?{query}')
+  })
+
+  it('redact digits', () => {
+    validate('/a/123', '/a/123')
+    validate('/a/12312bcdd23', '/a/:id')
+    validate('/a/B00I8BIC9E', '/a/:id')
+    // uuid
+    validate('/a/786c1883-dc0b-495f-a16b-6c53fb20b272', '/a/:id')
+  })
+
+  it('redact special characters', () => {
+    validate('/a-b-c-d', '/:id')
+    validate('/a~b-c', '/:id')
+    validate('/a~b-c/d/e/f', '/:id/*')
+    validate('/b%c%d-ef', '/:id')
+  })
+
+  it('react mix of lower-uppercase characters', () => {
+    validate('/abcDEF', '/:id')
+    validate('/a1W9FtW5DnkyP3Bucng4aLvqT/edit', '/:id/*')
+  })
+})
